+++ src/context/SignUpContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const SignUpContext = createContext(null);
