+++ src/pages/SignUp.jsx
@@ -1,12 +1,17 @@
 import Sidebar from "src/components/SignUpComponents/Sidebar";
 import Contents from "src/components/SignUpComponents/Contents";
 import { MainContainer } from "src/components/SignUpComponents/styled-components/SignUpStyle";
+import { useNavigate } from "react-router-dom";
+import { SignUpContext } from "src/context/SignUpContext";
 function SignUp() {
+  const navigate = useNavigate();
   return (
-    <MainContainer>
-      <Sidebar />
-      <Contents />
-    </MainContainer>
+    <SignUpContext.Provider value={navigate}>
+      <MainContainer>
+        <Sidebar />
+        <Contents />
+      </MainContainer>
+    </SignUpContext.Provider>
   );
 }
